perf(qa): isolate OTEL model-call test imports
